yolov5_garbage_detect-backend/requirements.txt
# YOLOv5 requirements

# Usage: pip install -r requirements.txt

# jupyter

# colored terminal output ----------------------------------------
colorama

# PyTorch ----------------------------------------
# pip install torch==1.13.1 torchvision==0.14.1 torchaudio==0.13.1

# Backend ----------------------------------------
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-wtf==1.1.1
flask_jwt_extended==4.4.4
flask-mail==0.9.1
email-validator==1.3.1
pymysql==1.0.2
cryptography==39.0.2

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
# torch>=1.7.0
# torchvision>=0.8.1
tqdm>=4.64.0
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb
# clearml

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=5.2  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export (or tensorflow-cpu, tensorflow-aarch64)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
